requirements.txt
# Medical Center Platform Migration Tool Dependencies
# Core database and data processing libraries

# Database drivers and ORM
PyMySQL>=1.1.0              # MySQL driver for Python
SQLAlchemy>=2.0.0            # Database toolkit and ORM
cryptography>=41.0.0         # Required for PyMySQL SSL connections

# Configuration and data formats
PyYAML>=6.0                  # YAML configuration file parsing
python-dotenv>=1.0.0         # Environment variable management

# Data processing and validation
pandas>=2.0.0                # Data manipulation and analysis
numpy>=1.24.0                # Numerical computing (required by pandas)

# Utilities and helpers
click>=8.1.0                 # Command line interface creation
rich>=13.0.0                 # Rich text and beautiful formatting in terminal
tqdm>=4.65.0                 # Progress bars for long-running operations
colorama>=0.4.6              # Cross-platform colored terminal text

# Logging and monitoring
structlog>=23.0.0            # Structured logging
prometheus-client>=0.17.0    # Metrics collection

# Testing and development (optional)
pytest>=7.4.0                # Testing framework
pytest-cov>=4.1.0            # Coverage reporting
pytest-mock>=3.11.0          # Mocking for tests
black>=23.0.0                # Code formatting
isort>=5.12.0                # Import sorting
flake8>=6.0.0                # Code linting
mypy>=1.5.0                  # Static type checking

# Performance and optimization
psutil>=5.9.0                # System and process utilities
memory-profiler>=0.61.0      # Memory usage profiling

# Security and encryption
cryptography>=41.0.0         # Cryptographic recipes and primitives
bcrypt>=4.0.0                # Password hashing

# Network and HTTP
requests>=2.31.0             # HTTP library for API calls
urllib3>=2.0.0               # HTTP client (required by requests)

# Date and time handling
python-dateutil>=2.8.0       # Date utilities
pytz>=2023.3                 # Timezone support

# File handling and compression
pathlib2>=2.3.7              # Object-oriented filesystem paths
zipfile36>=0.1.3             # ZIP file handling (Python 3.6 compatibility)

# Data serialization
msgpack>=1.0.5               # Fast binary serialization
ujson>=5.7.0                 # Fast JSON parsing and encoding

# Monitoring and observability
opentelemetry-api>=1.20.0    # OpenTelemetry API
opentelemetry-sdk>=1.20.0    # OpenTelemetry SDK
opentelemetry-instrumentation>=0.41b0  # Auto-instrumentation

# Error tracking and reporting
sentry-sdk>=1.28.0           # Error tracking and performance monitoring

# Configuration management
dynaconf>=3.2.0              # Dynamic configuration management
hydra-core>=1.3.0            # Configuration management framework

# Data validation
marshmallow>=3.20.0          # Object serialization/deserialization
pydantic>=2.0.0              # Data validation using Python type annotations
cerberus>=1.3.5               # Lightweight data validation library

# Database utilities
alembic>=1.11.0              # Database migration tool for SQLAlchemy
sqlacodegen>=3.0.0           # SQLAlchemy model generator

# Backup and restore
boto3>=1.28.0                # AWS SDK for Python (for S3 backups)
azure-storage-blob>=12.17.0  # Azure Blob Storage client
google-cloud-storage>=2.10.0  # Google Cloud Storage client

# Parallel processing
concurrent-futures>=3.1.1    # Backport of concurrent.futures (Python 2.7)
multiprocessing-logging>=0.3.0  # Multiprocessing logging support

# System integration
psutil>=5.9.0                # Cross-platform lib for process and system monitoring
pywin32>=306                 # Windows API access (Windows only)
pyobjc-framework-SystemConfiguration>=9.2  # macOS system configuration (macOS only)

# Development tools
pre-commit>=3.3.0            # Git hooks framework
tox>=4.6.0                   # Virtual environment management and testing
coverage>=7.2.0              # Code coverage measurement

# Documentation
sphinx>=7.0.0                # Documentation generator
sphinx-rtd-theme>=1.3.0      # Read the Docs theme for Sphinx
myst-parser>=2.0.0           # Markdown parser for Sphinx

# Performance profiling
cProfile>=0.0.0              # Python profiler
line-profiler>=4.0.0         # Line-by-line profiling
memory-profiler>=0.61.0      # Memory usage profiling

# Database performance
sqlalchemy-utils>=0.41.0     # SQLAlchemy utilities
sqlalchemy-migrate>=0.13.0    # Database schema migration

# Data export and import
openpyxl>=3.1.0              # Excel file handling
xlrd>=2.0.1                  # Excel file reading
tabulate>=0.9.0              # Pretty-print tabular data

# Network utilities
netaddr>=0.8.0               # Network address manipulation
dnspython>=2.4.0             # DNS toolkit

# Security utilities
passlib>=1.7.4               # Password hashing library
python-jose>=3.3.0           # JavaScript Object Signing and Encryption

# Monitoring and metrics
prometheus-client>=0.17.0    # Prometheus client library
statsd>=4.0.1                # Statsd client

# Error handling and debugging
better-exceptions>=0.3.0     # Better exception formatting
ipdb>=0.13.13                # IPython-enabled pdb

# Data visualization (optional)
matplotlib>=3.7.0            # Plotting library
seaborn>=0.12.0              # Statistical data visualization
plotly>=5.15.0               # Interactive plotting library

# Machine learning utilities (optional)
scikit-learn>=1.3.0          # Machine learning library
scipy>=1.11.0                # Scientific computing library

# Web framework (optional, for web-based monitoring)
flask>=2.3.0                 # Web framework
fastapi>=0.100.0             # Modern web framework for APIs

# Async support (optional)
asyncio-mqtt>=0.13.0         # Async MQTT client
aiofiles>=23.1.0             # Async file operations

# Cloud and container support
docker>=6.1.0                # Docker SDK for Python
kubernetes>=27.0.0            # Kubernetes Python client

# Backup and disaster recovery
rsync>=0.1.0                 # rsync wrapper for Python
tarfile>=0.0.0               # Tar file handling

# Data integrity and verification
hashlib>=0.0.0               # Hash algorithms
hmac>=0.0.0                  # HMAC implementation